Update Adapter Pattern : Observers (Event in AuctionCore ) -> Listeners (Event in System)

+++ server/src/main/java/server/service/listeners/AuctionEventListener.java
@@ -1,21 +1,23 @@
 package server.service.listeners;
 
+import java.math.BigDecimal;
+
 public interface AuctionEventListener {
-  void onOutbid(int userId, int auctionId, String itemName, double newPrice);
+  void onOutbid(int userId, int auctionId, String itemName, BigDecimal newPrice);
 
-  void onAuctionEnded(int winnerId, int auctionId, String itemName, double finalPrice);
+  void onAuctionEnded(int winnerId, int auctionId, String itemName, BigDecimal finalPrice);
 
-  void onBidPlaced(int bidderId, int auctionId, String itemName, double amount);
+  void onBidPlaced(int bidderId, int auctionId, String itemName, BigDecimal amount);
 
   void onAuctionStarted(int userId, int auctionId, String itemName);
 
-  void onAuctionWon(int winnerId, int auctionId, String itemName, double finalPrice);
+  void onAuctionWon(int winnerId, int auctionId, String itemName, BigDecimal finalPrice);
 
   void onAuctionLost(int loserId, int auctionId, String itemName);
 
-  void onPaymentReceived(int sellerId, int auctionId, String itemName, double amount);
+  void onPaymentReceived(int sellerId, int auctionId, String itemName, BigDecimal amount);
 
-  void onPaymentDue(int buyerId, int auctionId, String itemName, double amount);
+  void onPaymentDue(int buyerId, int auctionId, String itemName, BigDecimal amount);
 
   void onItemApproved(int sellerId, int itemId, String itemName);
