+++ server/src/main/java/server/service/listeners/NotificationEventHandler.java
@@ -3,6 +3,8 @@ package server.service.listeners;
 import server.service.NotificationService;
 import server.common.enums.NotificationType;
 
+import java.math.BigDecimal;
+
 public class NotificationEventHandler implements AuctionEventListener {
   private final NotificationService notificationService;
 
@@ -11,14 +13,14 @@ public class NotificationEventHandler implements AuctionEventListener {
   }
 
   @Override
-  public void onOutbid(int userId, int auctionId, String itemName, double newPrice) {
+  public void onOutbid(int userId, int auctionId, String itemName, BigDecimal newPrice) {
     notificationService.push(userId, "Outbid Alert!",
         "Someone bid higher ($" + newPrice + ") on [" + itemName + "].",
         NotificationType.OUTBID, auctionId);
   }
 
   @Override
-  public void onBidPlaced(int bidderId, int auctionId, String itemName, double amount) {
+  public void onBidPlaced(int bidderId, int auctionId, String itemName, BigDecimal amount) {
     notificationService.push(bidderId, "Bid Successful",
         "You placed a bid of $" + amount + " on [" + itemName + "].",
         NotificationType.BID_PLACED, auctionId);
@@ -32,14 +34,14 @@ public class NotificationEventHandler implements AuctionEventListener {
   }
 
   @Override
-  public void onAuctionEnded(int userId, int auctionId, String itemName, double finalPrice) {
+  public void onAuctionEnded(int userId, int auctionId, String itemName, BigDecimal finalPrice) {
     notificationService.push(userId, "Auction Ended",
         "Auction for [" + itemName + "] has closed at $" + finalPrice + ".",
         NotificationType.AUCTION_ENDED, auctionId);
   }
 
   @Override
-  public void onAuctionWon(int winnerId, int auctionId, String itemName, double finalPrice) {
+  public void onAuctionWon(int winnerId, int auctionId, String itemName, BigDecimal finalPrice) {
     notificationService.push(winnerId, "Congratulations!",
         "You won [" + itemName + "] for $" + finalPrice + "!",
         NotificationType.AUCTION_WON, auctionId);
@@ -53,7 +55,7 @@ public class NotificationEventHandler implements AuctionEventListener {
   }
 
   @Override
-  public void onPaymentDue(int buyerId, int auctionId, String itemName, double amount) {
+  public void onPaymentDue(int buyerId, int auctionId, String itemName, BigDecimal amount) {
     notificationService.push(buyerId, "Payment Required",
         "Please pay $" + amount + " for your won item: [" + itemName + "].",
 
@@ -61,7 +63,7 @@ public class NotificationEventHandler implements AuctionEventListener {
   }
 
   @Override
-  public void onPaymentReceived(int sellerId, int auctionId, String itemName, double amount) {
+  public void onPaymentReceived(int sellerId, int auctionId, String itemName, BigDecimal amount) {
     notificationService.push(sellerId, "Payment Received",
         "You received $" + amount + " for [" + itemName + "].",
         NotificationType.PAYMENT_RECEIVED, auctionId);
